Always export the data being displayed
When clicking on the export button and if the data cannot be fetched
from the current result handle, we run the query again to generate the
data to export.
However the query to run was the one being selected by the user, which
could be a different query from the one that was used to generate the
results being displayed.
Instead reuse the previous statement.